+++ petstore/src/main/java/petstore/api/spec/entities/Tag.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 @Data
 public class Tag {
-   private Integer id;
+   private Long id;
+   // private Integer id;
    private String name;
 }
